+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 ---
 ### Beta 0.3
 + Add changelog
++ Add navigation/title bar
 * Replace status bar button text with icons
 * Overhaul chat UI